Bump Python Versions (#241)
# Pull Request

## Description

This pull request updates several development dependencies in the
`pyproject.toml` files for the `scanner` and `tests` modules. The
changes ensure the use of newer versions of `pytest`, `ruff`, and `ty`.

Dependency updates:

*
[`scanner/pyproject.toml`](diffhunk://#diff-a3b82d49d76ebad61ba78436003c53e4907ae972476a9531807ceabd157d7586L14-R17):
Updated `pytest` to version `8.4.1`, `ruff` to version `0.12.1`, and
`ty` to version `0.0.1a12`.
*
[`tests/pyproject.toml`](diffhunk://#diff-03049af3cb225ee5c5f22bc10731df89191cc762ad9ca9aa4cb7d3a87786ac8dL7-R9):
Updated `pytest` to version `8.4.1`, `ruff` to version `0.12.1`, and
`ty` to version `0.0.1a12`.