Update Python Dependencies (#246)
# Pull Request

## Description

This pull request updates several dependencies in the `pyproject.toml`
files for `scanner` and `tests` to ensure compatibility and utilize the
latest versions. The changes include minor version bumps for tools used
in development and testing.

Dependency updates:

*
[`scanner/pyproject.toml`](diffhunk://#diff-a3b82d49d76ebad61ba78436003c53e4907ae972476a9531807ceabd157d7586L16-R22):
Updated `ruff` to version `0.12.2`, `ty` to version `0.0.1a14`, and `uv`
required version to `0.7.19`.
*
[`tests/pyproject.toml`](diffhunk://#diff-03049af3cb225ee5c5f22bc10731df89191cc762ad9ca9aa4cb7d3a87786ac8dL8-R13):
Updated `ruff` to version `0.12.2`, `ty` to version `0.0.1a14`, and `uv`
required version to `0.7.19`.