Fix Vite Build (#289)
# Pull Request

## Description

This pull request refactors the workflow for setting up and building the
dashboard in the deployment pipeline, simplifying the process by
consolidating steps into a reusable custom action. Additionally, it
introduces a `build` command in the `just` task runner for consistency
and maintainability.

### Workflow Refactoring:

*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3L80-R83):
Replaced multiple steps for setting up Node.js, installing dependencies,
and building the dashboard with a single reusable custom action,
`setup-dashboard-dependencies`, and updated the build command to use
`just dashboard::build` for consistency.

### Task Runner Update:

*
[`dashboard/dashboard.just`](diffhunk://#diff-bf57816edcdcaa85dfab81de274d3dba20e8e03fedb2008093bcb3903130f7aeR17-R20):
Added a `build` task to the `just` task runner to encapsulate the
production build process (`npm run build`).