Bump Python Dependencies (#302)
# Pull Request

## Description

This pull request updates several development dependencies in the
`pyproject.toml` files for both the `scanner` and `tests` directories.
The updates include minor version bumps for the `ruff` and `ty`
packages.

Dependency updates:

*
[`scanner/pyproject.toml`](diffhunk://#diff-a3b82d49d76ebad61ba78436003c53e4907ae972476a9531807ceabd157d7586L16-R17):
Updated `ruff` from `0.12.4` to `0.12.5` and `ty` from `0.0.1a15` to
`0.0.1a16`.
*
[`tests/pyproject.toml`](diffhunk://#diff-03049af3cb225ee5c5f22bc10731df89191cc762ad9ca9aa4cb7d3a87786ac8dL8-R9):
Updated `ruff` from `0.12.4` to `0.12.5` and `ty` from `0.0.1a15` to
`0.0.1a16`.